+++ pac4j-core/src/main/java/org/pac4j/core/profile/service/InMemoryProfileService.java
@@ -3,6 +3,7 @@ package org.pac4j.core.profile.service;
 import org.pac4j.core.profile.CommonProfile;
 import org.pac4j.core.profile.definition.CommonProfileDefinition;
 import org.pac4j.core.profile.factory.ProfileFactory;
+import org.pac4j.core.util.serializer.ProfileServiceSerializer;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -38,6 +39,7 @@ public class InMemoryProfileService<U extends CommonProfile> extends AbstractPro
         assertNotNull("passwordEncoder", getPasswordEncoder());
 
         defaultProfileDefinition(new CommonProfileDefinition(profileFactory));
+        setSerializer(new ProfileServiceSerializer(CommonProfile.class));
 
         super.internalInit();
     }