+++ pac4j-mongo/src/main/java/org/pac4j/mongo/profile/service/MongoProfileService.java
@@ -9,6 +9,7 @@ import org.pac4j.core.credentials.password.PasswordEncoder;
 import org.pac4j.core.profile.definition.CommonProfileDefinition;
 import org.pac4j.core.profile.service.AbstractProfileService;
 import org.pac4j.core.util.CommonHelper;
+import org.pac4j.core.util.serializer.ProfileServiceSerializer;
 import org.pac4j.mongo.profile.MongoProfile;
 
 import java.util.ArrayList;
@@ -59,7 +60,9 @@ public class MongoProfileService extends AbstractProfileService<MongoProfile> {
         CommonHelper.assertNotNull("mongoClient", this.mongoClient);
         CommonHelper.assertNotBlank("usersDatabase", this.usersDatabase);
         CommonHelper.assertNotBlank("usersCollection", this.usersCollection);
+
         defaultProfileDefinition(new CommonProfileDefinition(x -> new MongoProfile()));
+        setSerializer(new ProfileServiceSerializer(MongoProfile.class));
 
         super.internalInit();
     }